refactor: source code structural organization
