+++ docker-compose.yml
@@ -2,7 +2,7 @@ services:
   gotigram:
     image: tiagura/gotigram:latest
     container_name: gotigram
-    user: "${UID}:${GID}"
+    user: "${MYUID}:${MYGID}"
     # Option 1: Use external .env file
     # env_file:
     #  - .env
@@ -16,7 +16,8 @@ services:
       TELEGRAM_TOKEN: <TELEGRAM_TOKEN>
     # Optional: mount this volume to persist subscriptions outside container
     # Command /save won't work without this volume
-    # The startup import also requires this volume and to match the path defined in SUBSCRIPTIONS_FILE
+    # The startup import also requires this volume to match the path defined in SUBSCRIPTIONS_FILE
+      SUBSCRIPTIONS_FILE: /subscriptions/subscriptions.json
     volumes:
       - ./subscriptions:/subscriptions
     restart: unless-stopped
